+++ src/pages/forgotPassword.tsx
@@ -1,4 +1,4 @@
-import ForgotPassword from "@/components/forgot-password";
+import ForgotPassword from "@/components/forgotPassword";
 
 export default function ChangePasswordPage(): JSX.Element {
   return <ForgotPassword />;
